autocomplete search results working. Can't select a place yet.
